Check for streamURL in webcam settings
Insure that that there is a URL for the webcam stream in the returned JSON data from OctoPrint. OctoPrint 1.34 appears to return a webcam entry even though there is not a stream URL set in the settings. Without this check, the subsequent usage of stream_url will fail when streamURL is missing.